cleanup(gradle): move the safe dependency resolution into its own file
SafeTaskResolver, SafeResolveContext, the lock-free lookup and the
reflection readers behind them are one mechanism with one consumer,
resolveDependsOn. They now live in SafeTaskResolution.kt; TaskUtils.kt keeps
the target-building code and calls in. flattenDependsOn and flattenValues
become internal since both files use them.
